Session 5: FP64→FP32 conversions and comprehensive fix list
Successfully Converted to FP32 (+2 benchmarks):

1. degrid-sycl: Complete FP64→FP32 conversion
   - Changed Makefile: -DPRECISION=double → -DPRECISION=float
   - Added 'f' suffix to all float literals
   - ✓ Compiles and runs successfully on Arc A770

2. ddbp-sycl: Complete FP64→FP32 conversion
   - Replaced all double → float throughout codebase
   - Added 'f' suffix to all decimal literals (113 instances)
   - ✓ Compiles and runs successfully on Arc A770

Created Comprehensive Fix List:
- FIXME_LIST.md: Detailed roadmap to 100% working benchmarks
- Categorized all 61 remaining failures by fix type
- Provided automated scripts for batch fixes
- Estimated 11-17 hours to complete remaining work

Categories documented:
1. FP64→FP32 Conversions (~15-20 benchmarks)
2. Missing Headers/Files (~8-12 benchmarks)
3. SYCL API Issues (~5-8 benchmarks)
4. Makefile Problems (~3-5 benchmarks)
5. Namespace Issues (~2-5 benchmarks)
6. Complex Projects (~3-5 benchmarks)
7. Backend Issues (~1-3 benchmarks)

Priority phases defined:
- Phase 1: Quick wins (+10-15 benchmarks)
- Phase 2: FP32 conversions (+15-20 benchmarks)
- Phase 3: Dependencies (+3-5 benchmarks)
- Phase 4: SYCL API updates (+5-8 benchmarks)
- Phase 5: Complex cases (+3-5 benchmarks)

Overall Progress:
- Session 1-4: 376/444 → 381/444 (+5)
- Session 5: 381/444 → 383/444 (+2)
- Total: 383/444 (86.3%)
- Improvement: +127 benchmarks from starting point (+28.6%)

+++ ORG/SYCL/ddbp-sycl/main.cpp
@@ -50,7 +50,7 @@ using namespace sycl;
 
 void pad_projections_kernel(
     nd_item<3> &item,
-    double* d_img,
+    float* d_img,
     const int nDetXMap,
     const int nDetYMap,
     const int nElem,
@@ -63,11 +63,11 @@ void pad_projections_kernel(
 
 void map_boudaries_kernel(
     nd_item<3> &item,
-    double* d_pBound,
+    float* d_pBound,
     const int nElem,
-    const double valueLeftBound,
-    const double sizeElem,
-    const double offset)
+    const float valueLeftBound,
+    const float sizeElem,
+    const float offset)
 {
   const int gid = item.get_global_id(2);
   if (gid < nElem)
@@ -76,13 +76,13 @@ void map_boudaries_kernel(
 
 void rot_detector_kernel(
     nd_item<3> &item,
-          double* __restrict d_pRdetY,
-          double* __restrict d_pRdetZ,
-    const double* __restrict d_pYcoord,
-    const double* __restrict d_pZcoord,
-    const double yOffset,
-    const double zOffset,
-    const double phi,
+          float* __restrict d_pRdetY,
+          float* __restrict d_pRdetZ,
+    const float* __restrict d_pYcoord,
+    const float* __restrict d_pZcoord,
+    const float yOffset,
+    const float zOffset,
+    const float phi,
     const int nElem)
 {
   const int gid = item.get_global_id(2);
@@ -97,15 +97,15 @@ void rot_detector_kernel(
 
 void mapDet2Slice_kernel(
     nd_item<3> &item,
-           double* __restrict const pXmapp,
-           double* __restrict const pYmapp,
-    double tubeX,
-    double tubeY,
-    double tubeZ,
-    const double* __restrict const pXcoord,
-    const double* __restrict const pYcoord,
-    const double* __restrict const pZcoord,
-    const double* __restrict const pZSlicecoord,
+           float* __restrict const pXmapp,
+           float* __restrict const pYmapp,
+    float tubeX,
+    float tubeY,
+    float tubeZ,
+    const float* __restrict const pXcoord,
+    const float* __restrict const pYcoord,
+    const float* __restrict const pZcoord,
+    const float* __restrict const pZSlicecoord,
     const int nDetXMap,
     const int nDetYMap,
     const int nz)
@@ -128,7 +128,7 @@ void mapDet2Slice_kernel(
 
 void img_integration_kernel(
     nd_item<3> &item,
-    double* d_img,
+    float* d_img,
     const int nPixX,
     const int nPixY,
     const bool direction,
@@ -158,7 +158,7 @@ void img_integration_kernel(
 
       int spot = ty - s;
 
-      double val = 0;
+      float val = 0;
 
       if (spot >= 0) {
         val = d_img[(pz*nPixY*nPixX) + (offsetY + spot) * nPixY + px];
@@ -175,7 +175,7 @@ void img_integration_kernel(
 
       int spot = tx - s;
 
-      double val = 0;
+      float val = 0;
 
       if (spot >= 0) {
         val = d_img[(pz*nPixY*nPixX) + py * nPixY + spot + offsetX];
@@ -190,12 +190,12 @@ void img_integration_kernel(
 
 void bilinear_interpolation_kernel(
     nd_item<3> &item,
-          double* __restrict d_sliceI,
-    const double* __restrict d_pProj,
-    const double* __restrict d_pObjX,
-    const double* __restrict d_pObjY,
-    const double* __restrict d_pDetmX,
-    const double* __restrict d_pDetmY,
+          float* __restrict d_sliceI,
+    const float* __restrict d_pProj,
+    const float* __restrict d_pObjX,
+    const float* __restrict d_pObjY,
+    const float* __restrict d_pDetmX,
+    const float* __restrict d_pDetmY,
     const int nPixXMap,
     const int nPixYMap,
     const int nDetXMap,
@@ -215,62 +215,62 @@ void bilinear_interpolation_kernel(
 
   // Adjust the mapped coordinates to cross the range of (0-nDetX).*duMap 
   // Divide by pixelSize to get a unitary pixel size
-  const double xNormData = nDetX - d_pObjX[py] / d_pDetmX[0];
+  const float xNormData = nDetX - d_pObjX[py] / d_pDetmX[0];
   const int    xData = sycl::floor(xNormData);
-  const double alpha = xNormData - xData;
+  const float alpha = xNormData - xData;
 
   // Adjust the mapped coordinates to cross the range of (0-nDetY).*dyMap  
   // Divide by pixelSize to get a unitary pixel size
-  const double yNormData = (d_pObjY[px] / d_pDetmX[0]) - (d_pDetmY[0] / d_pDetmX[0]);
+  const float yNormData = (d_pObjY[px] / d_pDetmX[0]) - (d_pDetmY[0] / d_pDetmX[0]);
   const int    yData = sycl::floor(yNormData);
-  const double beta = yNormData - yData;
+  const float beta = yNormData - yData;
 
-  double d00, d01, d10, d11;
+  float d00, d01, d10, d11;
   if (((xNormData) >= 0) && ((xNormData) <= nDetX) && ((yNormData) >= 0) && ((yNormData) <= nDetY)) 
     d00 = d_pProj[(np*nDetYMap*nDetXMap) + (xData*nDetYMap + yData)];
   else
-    d00 = 0.0;
+    d00 = 0.0f;
 
   if (((xData + 1) > 0) && ((xData + 1) <= nDetX) && ((yNormData) >= 0) && ((yNormData) <= nDetY))
     d10 = d_pProj[(np*nDetYMap*nDetXMap) + ((xData + 1)*nDetYMap + yData)];
   else
-    d10 = 0.0;
+    d10 = 0.0f;
 
   if (((xNormData) >= 0) && ((xNormData) <= nDetX) && ((yData + 1) > 0) && ((yData + 1) <= nDetY))
     d01 = d_pProj[(np*nDetYMap*nDetXMap) + (xData*nDetYMap + yData + 1)];
   else
-    d01 = 0.0;
+    d01 = 0.0f;
 
   if (((xData + 1) > 0) && ((xData + 1) <= nDetX) && ((yData + 1) > 0) && ((yData + 1) <= nDetY))
     d11 = d_pProj[(np*nDetYMap*nDetXMap) + ((xData + 1)*nDetYMap + yData + 1)];
   else
-    d11 = 0.0;
+    d11 = 0.0f;
 
-  double result_temp1 = alpha * d10 + (-d00 * alpha + d00);
-  double result_temp2 = alpha * d11 + (-d01 * alpha + d01);
+  float result_temp1 = alpha * d10 + (-d00 * alpha + d00);
+  float result_temp2 = alpha * d11 + (-d01 * alpha + d01);
 
   d_sliceI[py * nPixYMap + px] = beta * result_temp2 + (-result_temp1 * beta + result_temp1);
 }
 
 void differentiation_kernel(
     nd_item<3> &item,
-          double* __restrict d_pVolume,
-    const double* __restrict d_sliceI,
-    double tubeX,
-    double rtubeY,
-    double rtubeZ,
-    const double* __restrict const d_pObjX,
-    const double* __restrict const d_pObjY,
-    const double* __restrict const d_pObjZ,
+          float* __restrict d_pVolume,
+    const float* __restrict d_sliceI,
+    float tubeX,
+    float rtubeY,
+    float rtubeZ,
+    const float* __restrict const d_pObjX,
+    const float* __restrict const d_pObjY,
+    const float* __restrict const d_pObjZ,
     const int nPixX,
     const int nPixY,
     const int nPixXMap,
     const int nPixYMap,
-    const double du,
-    const double dv,
-    const double dx,
-    const double dy,
-    const double dz,
+    const float du,
+    const float dv,
+    const float dx,
+    const float dy,
+    const float dz,
     const int nz) 
 {
   const int px = item.get_global_id(2);
@@ -313,12 +313,12 @@ void differentiation_kernel(
     int coordD = coordB + 1;
 
     // x - ray angle in X coord
-    double gamma = sycl::atan((d_pObjX[py] + (dx / 2.0) - tubeX) / (rtubeZ - d_pObjZ[nz]));
+    float gamma = sycl::atan((d_pObjX[py] + (dx / 2.0f) - tubeX) / (rtubeZ - d_pObjZ[nz]));
 
     // x - ray angle in Y coord
-    double alpha = sycl::atan((d_pObjY[px] + (dy / 2.0) - rtubeY) / (rtubeZ - d_pObjZ[nz]));
+    float alpha = sycl::atan((d_pObjY[px] + (dy / 2.0f) - rtubeY) / (rtubeZ - d_pObjZ[nz]));
 
-    double dA, dB, dC, dD;
+    float dA, dB, dC, dD;
 
     dA = d_sliceI[coordA];
     dB = d_sliceI[coordB];
@@ -339,7 +339,7 @@ void differentiation_kernel(
 
 void division_kernel(
     nd_item<3> &item,
-    double* d_img,
+    float* d_img,
     const int nPixX,
     const int nPixY,
     const int nSlices,
@@ -350,16 +350,16 @@ void division_kernel(
   const int pz = item.get_global_id(0);
   if (px < nPixY && py < nPixX && pz < nSlices) {
     const int pos = (nPixX*nPixY*pz) + (py * nPixY) + px;
-    d_img[pos] /= (double) nProj;
+    d_img[pos] /= (float) nProj;
   }
 }
 
 // Branchless distance-driven backprojection 
 void backprojectionDDb(
-          double* const h_pVolume,
-    const double* const h_pProj,
-    const double* const h_pTubeAngle,
-    const double* const h_pDetAngle,
+          float* const h_pVolume,
+    const float* const h_pProj,
+    const float* const h_pTubeAngle,
+    const float* const h_pDetAngle,
     const int idXProj,
     const int nProj,
     const int nPixX,
@@ -367,14 +367,14 @@ void backprojectionDDb(
     const int nSlices,
     const int nDetX,
     const int nDetY,
-    const double dx,
-    const double dy,
-    const double dz,
-    const double du,
-    const double dv,
-    const double DSD,
-    const double DDR,
-    const double DAG)
+    const float dx,
+    const float dy,
+    const float dz,
+    const float du,
+    const float dv,
+    const float DSD,
+    const float DDR,
+    const float DAG)
 {
   // Number of mapped detectors
   const int nDetXMap = nDetX + 1;
@@ -390,25 +390,25 @@ void backprojectionDDb(
   queue q(cpu_selector_v, property::queue::in_order());
 #endif
 
-  double *d_pProj = malloc_device<double>(nDetXMap*nDetYMap*nProj, q);
-  double *d_sliceI = malloc_device<double>(nPixXMap*nPixYMap, q);
-  double *d_pVolume = malloc_device<double>(nPixX*nPixY*nSlices, q);
+  float *d_pProj = malloc_device<float>(nDetXMap*nDetYMap*nProj, q);
+  float *d_sliceI = malloc_device<float>(nPixXMap*nPixYMap, q);
+  float *d_pVolume = malloc_device<float>(nPixX*nPixY*nSlices, q);
 
   // device memory for projections coordinates
-  double *d_pDetX = malloc_device<double>( nDetXMap , q);
-  double *d_pDetY = malloc_device<double>( nDetYMap , q);
-  double *d_pDetZ = malloc_device<double>( nDetYMap , q);
-  double *d_pObjX = malloc_device<double>( nPixXMap , q);
-  double *d_pObjY = malloc_device<double>( nPixYMap , q);
-  double *d_pObjZ = malloc_device<double>( nSlices , q);
+  float *d_pDetX = malloc_device<float>( nDetXMap , q);
+  float *d_pDetY = malloc_device<float>( nDetYMap , q);
+  float *d_pDetZ = malloc_device<float>( nDetYMap , q);
+  float *d_pObjX = malloc_device<float>( nPixXMap , q);
+  float *d_pObjY = malloc_device<float>( nPixYMap , q);
+  float *d_pObjZ = malloc_device<float>( nSlices , q);
 
   // device memory for mapped coordinates
-  double *d_pDetmY = malloc_device<double>( nDetYMap , q);
-  double *d_pDetmX = malloc_device<double>( nDetYMap * nDetXMap , q);
+  float *d_pDetmY = malloc_device<float>( nDetYMap , q);
+  float *d_pDetmX = malloc_device<float>( nDetYMap * nDetXMap , q);
 
   // device memory for rotated detector coords
-  double *d_pRdetY = malloc_device<double>( nDetYMap , q);
-  double *d_pRdetZ = malloc_device<double>( nDetYMap , q);
+  float *d_pRdetY = malloc_device<float>( nDetYMap , q);
+  float *d_pRdetZ = malloc_device<float>( nDetYMap , q);
 
   auto start = std::chrono::steady_clock::now();
 
@@ -421,8 +421,8 @@ void backprojectionDDb(
   // Copy projection data padding with zeros for image integation
 
   // Initialize first column and row with zeros
-  const double* h_pProj_tmp;
-  double* d_pProj_tmp;
+  const float* h_pProj_tmp;
+  float* d_pProj_tmp;
 
   lws[2] = maxThreadsPerBlock;
   gws[2] = (nDetXMap / maxThreadsPerBlock + 1) * maxThreadsPerBlock;
@@ -439,7 +439,7 @@ void backprojectionDDb(
 
     // Pad on Y coord direction
     d_pProj_tmp = d_pProj + (nDetXMap*nDetYMap*np) + 1;
-    q.memset(d_pProj_tmp, 0, nPixY * sizeof(double));
+    q.memset(d_pProj_tmp, 0, nPixY * sizeof(float));
   }
 
   // Copy projections data from host to device
@@ -447,7 +447,7 @@ void backprojectionDDb(
     for (int c = 0; c < nDetX; c++) {
       h_pProj_tmp = h_pProj + (c * nDetY) + (nDetX*nDetY*np);
       d_pProj_tmp = d_pProj + (((c + 1) * nDetYMap) + 1) + (nDetXMap*nDetYMap*np);
-      q.memcpy(d_pProj_tmp, h_pProj_tmp, nDetY * sizeof(double));
+      q.memcpy(d_pProj_tmp, h_pProj_tmp, nDetY * sizeof(float));
     }
 
   // Generate detector and object boudaries
@@ -458,7 +458,7 @@ void backprojectionDDb(
 
   q.submit([&] (handler &cgh) {
     cgh.parallel_for<class map_detX>(nd_range<3>(gws, lws), [=] (nd_item<3> item) {
-      map_boudaries_kernel(item, d_pDetX, nDetXMap, (double)nDetX, -du, 0.0);
+      map_boudaries_kernel(item, d_pDetX, nDetXMap, (float)nDetX, -du, 0.0f);
     });
   });
 
@@ -466,7 +466,7 @@ void backprojectionDDb(
 
   q.submit([&] (handler &cgh) {
     cgh.parallel_for<class map_detY>(nd_range<3>(gws, lws), [=] (nd_item<3> item) {
-      map_boudaries_kernel(item, d_pDetY, nDetYMap, nDetY / 2.0, dv, 0.0);
+      map_boudaries_kernel(item, d_pDetY, nDetYMap, nDetY / 2.0f, dv, 0.0f);
     });
   });
 
@@ -474,7 +474,7 @@ void backprojectionDDb(
 
   q.submit([&] (handler &cgh) {
     cgh.parallel_for<class map_pixX>(nd_range<3>(gws, lws), [=] (nd_item<3> item) {
-      map_boudaries_kernel(item, d_pObjX, nPixXMap, (double)nPixX, -dx, 0.0);
+      map_boudaries_kernel(item, d_pObjX, nPixXMap, (float)nPixX, -dx, 0.0f);
     });
   });
 
@@ -482,7 +482,7 @@ void backprojectionDDb(
 
   q.submit([&] (handler &cgh) {
     cgh.parallel_for<class map_pixY>(nd_range<3>(gws, lws), [=] (nd_item<3> item) {
-      map_boudaries_kernel(item, d_pObjY, nPixYMap, nPixY / 2.0, dy, 0.0);
+      map_boudaries_kernel(item, d_pObjY, nPixYMap, nPixY / 2.0f, dy, 0.0f);
     });
   });
 
@@ -490,22 +490,22 @@ void backprojectionDDb(
 
   q.submit([&] (handler &cgh) {
     cgh.parallel_for<class map_pixZ>(nd_range<3>(gws, lws), [=] (nd_item<3> item) {
-      map_boudaries_kernel(item, d_pObjZ, nSlices, 0.0, dz, DAG + (dz / 2.0));
+      map_boudaries_kernel(item, d_pObjZ, nSlices, 0.0f, dz, DAG + (dz / 2.0f));
     });
   });
 
   // Initiate variables value with 0
-  q.memset(d_pDetZ, 0, nDetYMap * sizeof(double));
-  q.memset(d_pVolume, 0, nPixX * nPixY * nSlices * sizeof(double));
+  q.memset(d_pDetZ, 0, nDetYMap * sizeof(float));
+  q.memset(d_pVolume, 0, nPixX * nPixY * nSlices * sizeof(float));
 
   // X - ray tube initial position
-  double tubeX = 0;
-  double tubeY = 0;
-  double tubeZ = DSD;
+  float tubeX = 0;
+  float tubeY = 0;
+  float tubeZ = DSD;
 
   // Iso - center position
-  double isoY = 0;
-  double isoZ = DDR;
+  float isoY = 0;
+  float isoZ = DDR;
 
   // Integration of 2D projection over the whole projections
   // (S.1.Integration. - Liu et al(2017))
@@ -567,16 +567,16 @@ void backprojectionDDb(
   for (int p = projIni; p < projEnd; p++) {
 
     // Get specif tube angle for the projection
-    double theta = h_pTubeAngle[p] * M_PI / 180.0;
+    float theta = h_pTubeAngle[p] * M_PI / 180.0f;
 
     // Get specif detector angle for the projection
-    double phi = h_pDetAngle[p] * M_PI / 180.0;
+    float phi = h_pDetAngle[p] * M_PI / 180.0f;
 
     //printf("Tube angle:%f Det angle:%f\n", theta, phi);
 
     // Tube rotation
-    double rtubeY = ((tubeY - isoY)*std::cos(theta) - (tubeZ - isoZ)*std::sin(theta)) + isoY;
-    double rtubeZ = ((tubeY - isoY)*std::sin(theta) + (tubeZ - isoZ)*std::cos(theta)) + isoZ;
+    float rtubeY = ((tubeY - isoY)*std::cos(theta) - (tubeZ - isoZ)*std::sin(theta)) + isoY;
+    float rtubeZ = ((tubeY - isoY)*std::sin(theta) + (tubeZ - isoZ)*std::cos(theta)) + isoZ;
 
     //printf("R tube Y:%f R tube Z:%f\n", rtubeY, rtubeZ);
 
@@ -674,7 +674,7 @@ void backprojectionDDb(
   auto time = std::chrono::duration_cast<std::chrono::nanoseconds>(end - start).count();
   printf("Total kernel execution time %f (s)\n", time * 1e-9f);
 
-  q.memcpy(h_pVolume, d_pVolume, nSlices* nPixX * nPixY * sizeof(double));
+  q.memcpy(h_pVolume, d_pVolume, nSlices* nPixX * nPixY * sizeof(float));
 
   free(d_pProj, q);
   free(d_sliceI, q);
@@ -705,40 +705,40 @@ int main()
   const int nProj = 15;     // number of projections
   const int idXProj = -1;   // loop over all projections
 
-  const double dx = 0.112;  // single voxel size (mm)
-  const double dy = 0.112;
-  const double dz = 1.0;
+  const float dx = 0.112f;  // single voxel size (mm)
+  const float dy = 0.112f;
+  const float dz = 1.0f;
 
-  const double du = 0.14;   // single detector size (mm)
-  const double dv = 0.14;
+  const float du = 0.14f;   // single detector size (mm)
+  const float dv = 0.14f;
 
-  const double DSD = 700;   // distance from source to detector (mm)
-  const double DDR = 0.0;   // distance from detector to pivot (mm)
-  const double DAG = 25.0;  // distance of air gap (mm)
+  const float DSD = 700;   // distance from source to detector (mm)
+  const float DDR = 0.0f;   // distance from detector to pivot (mm)
+  const float DAG = 25.0f;  // distance of air gap (mm)
 
   const size_t pixVol = nPixX * nPixY * nSlices;
   const size_t detVol = nDetX * nDetY * nProj;
-  double *h_pVolume = (double*) malloc (pixVol * sizeof(double));
-  double *h_pProj = (double*) malloc (detVol * sizeof(double));
+  float *h_pVolume = (float*) malloc (pixVol * sizeof(float));
+  float *h_pProj = (float*) malloc (detVol * sizeof(float));
 
-  double *h_pTubeAngle = (double*) malloc (nProj * sizeof(double));
-  double *h_pDetAngle = (double*) malloc (nProj * sizeof(double));
+  float *h_pTubeAngle = (float*) malloc (nProj * sizeof(float));
+  float *h_pDetAngle = (float*) malloc (nProj * sizeof(float));
   
   // tube angles in degrees
   for (int i = 0; i < nProj; i++) 
-    h_pTubeAngle[i] = -7.5 + i * 15.0/nProj;
+    h_pTubeAngle[i] = -7.5f + i * 15.0f/nProj;
 
   // detector angles in degrees
   for (int i = 0; i < nProj; i++) 
-    h_pDetAngle[i] = -2.1 + i * 4.2/nProj;
+    h_pDetAngle[i] = -2.1f + i * 4.2f/nProj;
 
   // random values
   srand(123);
   for (size_t i = 0; i < pixVol; i++) 
-    h_pVolume[i] = (double)rand() / (double)RAND_MAX;
+    h_pVolume[i] = (float)rand() / (float)RAND_MAX;
 
   for (size_t i = 0; i < detVol; i++) 
-    h_pProj[i] = (double)rand() / (double)RAND_MAX;
+    h_pProj[i] = (float)rand() / (float)RAND_MAX;
 
   backprojectionDDb(
     h_pVolume,
@@ -754,7 +754,7 @@ int main()
     du, dv,
     DSD, DDR, DAG);
 
-  double checkSum = 0;
+  float checkSum = 0;
   for (size_t i = 0; i < pixVol; i++)
     checkSum += h_pVolume[i];
   printf("checksum = %lf\n", checkSum);
